Add format and format-all tasks to grunt
`grunt format` now runs js-beautify and sortImports on all js files in a given repo (or the current one if none is given). In addition, you can use `grunt format --verify` to just check that the files are formatted, but not actually modify them. This could be useful for a pre-commit hook.

The `grunt format-all` variation is similar to lint-all in that it runs format on all the repos that are libraries of the supplied repo.